requirements-frozen.txt
boto3==1.15.16
botocore==1.18.16
crhelper==2.0.6
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
python-dateutil==2.8.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
